Core/Script: fix build.
